samples/OverwatchLocalMCP/pyproject.toml
[project]
name = "overwatch-local-maestro-action"
version = "0.0.2.4"
description = "Overwatch Local MCP Server Agent for Instance Control. Also includes Action center MCP servers"
authors = [{ name = "UiPath", email = "support@uipath.com" }]
dependencies = [
    "langgraph>=0.2.70",
    "langchain-core>=0.3.34",
    "langchain-anthropic>=0.3.9",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "python-dotenv>=1.0.1",
    "uipath-langchain==0.0.113",
    "pydantic>=2.10.6",
    "aiohttp>=3.11.12",
    "typing-extensions>=4.12.2",
    "langchain-mcp-adapters>=0.1.0",
    "ipython>=8.32.0",
    "mcp>=1.4.1",
    "httpx>=0.27.0",
]
requires-python = ">=3.10" 
